+++ template/src/bin/main.rs
@@ -1,6 +1,11 @@
 //INCLUDEFILE !embassy
 #![no_std]
 #![no_main]
+#![deny(
+    clippy::mem_forget,
+    reason = "mem::forget is generally not safe to do with esp_hal types, especially those \
+    holding buffers for the duration of a data transfer."
+)]
 
 use esp_hal::{
     clock::CpuClock,